+++ app/src/test/java/com/ebsolutions/papertrail/financialdataingestionservice/ActuatorSteps.java
@@ -5,7 +5,7 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import com.ebsolutions.papertrail.financialdataingestionservice.config.TestConstants;
+import com.ebsolutions.papertrail.financialdataingestionservice.config.UriConstants;
 import com.ebsolutions.papertrail.financialdataingestionservice.tooling.BaseTest;
 import com.jayway.jsonpath.DocumentContext;
 import com.jayway.jsonpath.JsonPath;
@@ -23,14 +23,14 @@ public class ActuatorSteps extends BaseTest {
 
   @Given("application is up")
   public void applicationIsUp() throws Exception {
-    mockMvc.perform(get(TestConstants.HEALTH_CHECK_URI))
+    mockMvc.perform(get(UriConstants.HEALTH_CHECK_URI))
         .andExpect(status().isOk())
         .andExpect(jsonPath("$.status", is("UP")));
   }
 
   @When("the info endpoint is invoked")
   public void theInfoEndpointIsInvoked() throws Exception {
-    result = mockMvc.perform(get(TestConstants.INFO_CHECK_URI)).andReturn();
+    result = mockMvc.perform(get(UriConstants.INFO_CHECK_URI)).andReturn();
   }
 
   @Then("the correct info response is returned")