+++ app/src/test/java/com/ebsolutions/papertrail/financialdataingestionservice/config/UriConstants.java
@@ -1,7 +1,8 @@
 package com.ebsolutions.papertrail.financialdataingestionservice.config;
 
-public class TestConstants {
+public class UriConstants {
   // URIs
   public static final String HEALTH_CHECK_URI = "/actuator/health";
   public static final String INFO_CHECK_URI = "/actuator/info";
+  public static final String ACCOUNT_TRANSACTIONS_URI = "/account-transactions";
 }